[PI-259] Allow zip generation

+++ scripts/compress.js
@@ -1,14 +1,16 @@
 const fs = require('fs');
 const path = require('path');
 const ChromeExtension = require('crx');
+const argv = require('minimist')(process.argv.slice(2));
 /* eslint import/no-unresolved: 0 */
 const name = require('../build/manifest.json').name;
-const argv = require('minimist')(process.argv.slice(2));
 
-const keyPath = argv.key || 'key.pem';
+const keyPath = argv.key;
 const existsKey = fs.existsSync(keyPath);
+const zipOnly = argv['zip-only'];
+const isCrx = !zipOnly;
 
-if (!argv.codebase || !existsKey) {
+if (!argv.codebase || (isCrx && !existsKey)) {
   console.error('Missing input data.');
   return;
 }
@@ -16,21 +18,22 @@ if (!argv.codebase || !existsKey) {
 const crx = new ChromeExtension({
   appId: argv['app-id'],
   codebase: argv.codebase,
-  privateKey: existsKey ? fs.readFileSync(keyPath) : null
+  privateKey: existsKey
+    ? fs.readFileSync(keyPath)
+    : null
 });
 
-crx
-  .load(path.join(__dirname, '../build'))
-  .then(() => crx.loadContents())
-  .then(archiveBuffer => {
-    fs.writeFileSync(`${name}.zip`, archiveBuffer);
-    return crx.pack(archiveBuffer);
-  })
-  .then(crxBuffer => {
+async function compress(isCrxBuild) {
+  await crx.load(path.join(__dirname, '../build'));
+  const archiveBuffer = await crx.loadContents();
+  fs.writeFileSync(`${name}.zip`, archiveBuffer);
+  if (isCrxBuild) {
+    const crxBuffer = await crx.pack(archiveBuffer);
     const updateXML = crx.generateUpdateXML();
     fs.writeFileSync('update.xml', updateXML);
     fs.writeFileSync(`${name}-${argv.env}.crx`, crxBuffer);
     fs.unlinkSync(`${name}.zip`);
-    return;
-  })
-  .catch(err => console.error(err));
+  }
+}
+
+compress(isCrx).catch(err => console.error(err));
